src/pyproject.toml
[project]
name = "async-await"
version = "0.1.0"
description = "Play around with async await"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.10.5",
    "asyncio>=3.4.3",
    "bs4>=0.0.2",
    "colorama>=0.4.6",
    "httpx>=0.27.2",
    "requests>=2.32.3",
    "ruff>=0.6.4",
    "fastapi[standard]",
    "twirp",
    "redis[hiredis]",
    "prometheus_client>=0.21.0",
    "prometheus-fastapi-instrumentator",
    "anyio[trio]",
    "numpy",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpc-interceptor>=0.15.4",
]

[project.optional-dependencies]
build = ["setuptools", "cython"]
compile = ["cchardet"]

[tool.uv]
environments = [
    "sys_platform == 'darwin'",
    "sys_platform == 'linux'",
]
